+++ python/fastfoodordering/core/web/tool_calling/dspy_tools.py
@@ -79,7 +79,8 @@ class DSPyBrowserToolCaller(DSPyToolCaller, BrowserToolCaller):
         subtask: str,
         previous_browser_screenshot: Optional[cv2.typing.MatLike] = None,
         current_browser_screenshot: Optional[cv2.typing.MatLike] = None,
-        current_browser_snapshot: Optional[str] = None
+        current_browser_snapshot: Optional[str] = None,
+        banned_tools: Optional[str] = None,
     ) -> GeneratedTool:
         return await super().determine_tool(
             overall_goal=overall_goal,
@@ -88,6 +89,7 @@ class DSPyBrowserToolCaller(DSPyToolCaller, BrowserToolCaller):
             current_browser_snapshot=current_browser_snapshot, # if visibility
             previous_browser_screenshot=dspy_image_or_blank(previous_browser_screenshot), # if visibility
             current_browser_screenshot=dspy_image_or_blank(current_browser_screenshot), # if visibility
+            banned_tools=banned_tools,
         )
     
     async def determine_and_call_tools(
@@ -115,7 +117,8 @@ class DSPyBrowserToolCaller(DSPyToolCaller, BrowserToolCaller):
             if screenshot_filepath is None:
                 return None
             image = cv2.imread(screenshot_filepath)
-            if os.path.exists(screenshot_filepath) and \
+            if screenshot_filepath != os.getenv("CURRENT_SCREENSHOT_PATH", None) and \
+                os.path.exists(screenshot_filepath) and \
                 os.path.isfile(screenshot_filepath) and \
                 os.path.splitext(screenshot_filepath)[-1] in ["jpg", "jpeg", "png"]:
                     os.remove(screenshot_filepath)
